% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server_headers.R
\name{get_js_resources}
\alias{get_js_resources}
\title{Retrieves HTML attributes of any script tags to check
subresource integrity}
\usage{
get_js_resources(res)
}
\description{
Retrieves HTML attributes of any script tags to check
subresource integrity
}
