% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_summary-referrer-policy.R,
%   R/server_header_summaries.R
\name{header_summary.referrer-policy}
\alias{header_summary.referrer-policy}
\alias{header_summary}
\alias{header_summary.default}
\alias{header_summary.scheme}
\alias{header_summary.content-type}
\alias{header_summary.report-to}
\alias{header_summary.nel}
\alias{header_summary.content-security-policy-report-only}
\alias{header_summary.content-security-policy}
\alias{header_summary.access-control-allow-origin}
\alias{header_summary.strict-transport-security}
\alias{header_summary.x-frame-options}
\alias{header_summary.x-content-type-options}
\alias{header_summary.permissions-policy}
\alias{header_summary.server}
\alias{header_summary.x-xss-protection}
\alias{header_summary.x-permitted-cross-domain-policies}
\alias{header_summary.x-powered-by}
\alias{header_summary.cookies}
\alias{header_summary.redirection}
\alias{header_summary.subresource-integrity}
\title{Header Summaries}
\usage{
\method{header_summary}{`referrer-policy`}(value, ...)

header_summary(value, ...)

\method{header_summary}{default}(value, ...)

\method{header_summary}{scheme}(value, ...)

\method{header_summary}{`content-type`}(value, ...)

\method{header_summary}{`report-to`}(value, ...)

\method{header_summary}{nel}(value, ...)

\method{header_summary}{`content-security-policy-report-only`}(value, ...)

\method{header_summary}{`content-security-policy`}(value, ...)

\method{header_summary}{`access-control-allow-origin`}(value, ...)

\method{header_summary}{`strict-transport-security`}(value, ...)

\method{header_summary}{`x-frame-options`}(value, ...)

\method{header_summary}{`x-content-type-options`}(value, ...)

\method{header_summary}{`permissions-policy`}(value, ...)

\method{header_summary}{server}(value, ...)

\method{header_summary}{`x-xss-protection`}(value, ...)

\method{header_summary}{`x-permitted-cross-domain-policies`}(value, ...)

\method{header_summary}{`x-powered-by`}(value, ...)

\method{header_summary}{cookies}(value, ...)

\method{header_summary}{redirection}(value, ...)

\method{header_summary}{`subresource-integrity`}(value, ...)
}
\arguments{
\item{value}{Header element, likely obtained from httr::headers}

\item{...}{Other values (not used)}
}
\value{
A tibble or NULL (where the server element isn't known)
}
\description{
Summaries server header elements.
}
