% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server_headers.R
\name{is_redirected}
\alias{is_redirected}
\title{Is website redirection implemented i.e. HTTP -> HTTPS?}
\usage{
is_redirected(res)
}
\description{
Is website redirection implemented i.e. HTTP -> HTTPS?
}
